@@ -1,20 +1,46 @@
 <template>
-    <div class="w-full flex flex-col gap-2">
-        <CapabilitiesEnvironmentCharts v-for="environment in environments" :environment="environment"
-            :aspect="settings_selectedAspect"></CapabilitiesEnvironmentCharts>
+  <div class="w-full flex flex-col gap-2">
+    <CapabilitiesEnvironmentCharts
+      v-for="environment in environments"
+      :key="environment.id"
+      :environment="environment"
+      :aspect="selectedAspect"
+    />
+    <div v-if="!environments?.length" class="text-gray-500 text-center py-4">
+      {{ strings?.no_environments_message }}
     </div>
+  </div>
 </template>
-  
-<script>
-import { mapState } from 'pinia'
-import store from '../../../Store';
-import CapabilitiesEnvironmentCharts from './CapabilitiesEnvironmentCharts.vue';
 
-export default {
-    components: { CapabilitiesEnvironmentCharts },
-    computed: {
-        ...mapState(store, ["strings", "settings_selectedAspect", "environments"]),
-    },
+<script setup>
+import { computed, watch } from "vue";
+import { storeToRefs } from "pinia";
+import { useStore } from "../../../stores/index";
+import { useEnvironmentStore } from "../../../stores/environmentStore";
+import { useImpactStore } from "../../../stores/impactStore";
+import CapabilitiesEnvironmentCharts from "./CapabilitiesEnvironmentCharts.vue";
 
-}
+// Store setup
+const store = useStore();
+const environmentStore = useEnvironmentStore();
+const impactStore = useImpactStore();
+
+// Store refs
+const { strings } = storeToRefs(store);
+const { environments } = storeToRefs(environmentStore);
+const { settings_selectedAspect: selectedAspect } = storeToRefs(impactStore);
+
+// Optional: Validate environments data
+const validateEnvironments = computed(() => {
+  return environments.value?.every(
+    (env) => env.id && typeof env.id === "string"
+  );
+});
+
+// Optional: Watch for data issues
+watch(validateEnvironments, (isValid) => {
+  if (!isValid) {
+    console.warn("Invalid environment data detected");
+  }
+});
 </script>
