@@ -1,10 +1,7 @@
-
 export class Environment {
-    constructor({ id, display_name_en, display_name_fr }) {
-        this.id = id;
-        this.display_name_en = display_name_en;
-        this.display_name_fr = display_name_fr;
-    }
-
-
+  constructor({ id, display_name_en, display_name_fr }) {
+    this.id = id;
+    this.display_name_en = display_name_en;
+    this.display_name_fr = display_name_fr;
+  }
 }
