@@ -1,16 +1,15 @@
 export const DivergentColorPalette = [
-    "#2d5071",
-    "#d69e2e",
-    "#c85c7a",
-    "#dd6962",
-    "#a5598a",
-    "#7b598e",
-    "#e28146",
-    "#505685"
-]
+  "#2d5071",
+  "#d69e2e",
+  "#c85c7a",
+  "#dd6962",
+  "#a5598a",
+  "#7b598e",
+  "#e28146",
+  "#505685",
+];
 
 export function colorForIndex(index) {
-    let colorForIndex = DivergentColorPalette[index];
-    return colorForIndex;
+  let colorForIndex = DivergentColorPalette[index];
+  return colorForIndex;
 }
-
